Set project to use the C++20 language standard

+++ src/qt/bitcoingui.cpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2011-2019 The Bitcoin Core developers
-// Copyright (c) 2020-2023 The Bitcoin developers
+// Copyright (c) 2020-2025 The Bitcoin developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
@@ -247,7 +247,7 @@ void BitcoinGUI::createActions() {
     overviewAction->setStatusTip(tr("Show general overview of wallet"));
     overviewAction->setToolTip(overviewAction->statusTip());
     overviewAction->setCheckable(true);
-    overviewAction->setShortcut(QKeySequence(Qt::ALT + Qt::Key_1));
+    overviewAction->setShortcut(QKeySequence(Qt::ALT + static_cast<int>(Qt::Key_1)));
     tabGroup->addAction(overviewAction);
 
     sendCoinsAction = new QAction(
@@ -255,7 +255,7 @@ void BitcoinGUI::createActions() {
     sendCoinsAction->setStatusTip(tr("Send coins to a Bitcoin Cash address"));
     sendCoinsAction->setToolTip(sendCoinsAction->statusTip());
     sendCoinsAction->setCheckable(true);
-    sendCoinsAction->setShortcut(QKeySequence(Qt::ALT + Qt::Key_2));
+    sendCoinsAction->setShortcut(QKeySequence(Qt::ALT + static_cast<int>(Qt::Key_2)));
     tabGroup->addAction(sendCoinsAction);
 
     sendCoinsMenuAction =
@@ -273,7 +273,7 @@ void BitcoinGUI::createActions() {
                 config->GetChainParams().CashAddrPrefix())));
     receiveCoinsAction->setToolTip(receiveCoinsAction->statusTip());
     receiveCoinsAction->setCheckable(true);
-    receiveCoinsAction->setShortcut(QKeySequence(Qt::ALT + Qt::Key_3));
+    receiveCoinsAction->setShortcut(QKeySequence(Qt::ALT + static_cast<int>(Qt::Key_3)));
     tabGroup->addAction(receiveCoinsAction);
 
     receiveCoinsMenuAction =
@@ -288,7 +288,7 @@ void BitcoinGUI::createActions() {
     historyAction->setStatusTip(tr("Browse transaction history"));
     historyAction->setToolTip(historyAction->statusTip());
     historyAction->setCheckable(true);
-    historyAction->setShortcut(QKeySequence(Qt::ALT + Qt::Key_4));
+    historyAction->setShortcut(QKeySequence(Qt::ALT + static_cast<int>(Qt::Key_4)));
     tabGroup->addAction(historyAction);
 
 #ifdef ENABLE_WALLET
@@ -324,7 +324,7 @@ void BitcoinGUI::createActions() {
     quitAction = new QAction(platformStyle->TextColorIcon(":/icons/quit"),
                              tr("E&xit"), this);
     quitAction->setStatusTip(tr("Quit application"));
-    quitAction->setShortcut(QKeySequence(Qt::CTRL + Qt::Key_Q));
+    quitAction->setShortcut(QKeySequence(Qt::CTRL + static_cast<int>(Qt::Key_Q)));
     quitAction->setMenuRole(QAction::QuitRole);
     aboutAction = new QAction(platformStyle->TextColorIcon(":/icons/about"),
                               tr("&About %1").arg(PACKAGE_NAME), this);
@@ -490,7 +490,7 @@ void BitcoinGUI::createMenuBar() {
     QAction *minimize_action = window_menu->addAction(tr("Minimize"));
     minimize_action->setStatusTip(tr("Minimize the Main Window"));
     minimize_action->setToolTip(minimize_action->statusTip());
-    minimize_action->setShortcut(QKeySequence(Qt::CTRL + Qt::Key_M));
+    minimize_action->setShortcut(QKeySequence(Qt::CTRL + static_cast<int>(Qt::Key_M)));
     connect(minimize_action, &QAction::triggered,
             [] { QApplication::activeWindow()->showMinimized(); });
     connect(qApp, &QApplication::focusWindowChanged,
